[android] Moved navigation up listener to another package and removed redundant package 'activity'
